server/requirements.txt
numpy==1.19.5
gevent==21.1.2
seaborn==0.11.1
torch==1.8.1
torchvision==0.9.1
onnx==1.9.0
matplotlib==3.4.1
pandas==1.2.4
coremltools==4.1
thop==0.0.31.post2005241907
Flask==1.1.2
requests==2.25.1
scipy==1.6.2
tqdm==4.60.0
opencv_python==4.5.1.48
Pillow==8.2.0
PyYAML==5.4.1
wandb==0.10.31
